+++ app/volt/tasks/store_tasks.rb
@@ -1,7 +1,7 @@
 require 'mongo'
 require 'volt/models'
 
-class StoreTasks < Volt::TaskHandler
+class StoreTasks < Volt::Task
   def initialize(channel = nil, dispatcher = nil)
     @channel = channel
     @dispatcher = dispatcher
@@ -25,8 +25,8 @@ class StoreTasks < Volt::TaskHandler
       # Create a buffer
       buffer = model.buffer
 
-      # Assign the data
-      buffer.assign_attributes(data, true)
+      # Assign the changed data to the buffer
+      buffer.assign_attributes(data, false, true)
 
       buffer
     end
@@ -35,8 +35,12 @@ class StoreTasks < Volt::TaskHandler
   def save(collection, path, data)
     data = data.symbolize_keys
     promise = nil
-    Volt::Model.no_validate do
-      promise = load_model(collection, path, data)
+
+    # Don't check the permissions when we load the model, since we want all fields
+    Volt.skip_permissions do
+      Volt::Model.no_validate do
+        promise = load_model(collection, path, data)
+      end
     end
 
     # On the backend, the promise is resolved before its returned, so we can
@@ -55,14 +59,19 @@ class StoreTasks < Volt::TaskHandler
 
       Thread.current['in_channel'] = nil
 
-      puts "DONE--"
       next save_promise
     end
   end
 
   def delete(collection, id)
     # Load the model, then call .destroy on it
-    store.send(:"_#{collection}").where(_id: id).fetch_first do |model|
+    query = nil
+
+    Volt.skip_permissions do
+      query = store.send(:"_#{collection}").where(_id: id)
+    end
+
+    query.fetch_first do |model|
       if model
         if model.can_delete?
           db[collection].remove('_id' => id)