+++ lib/volt/models/model.rb
@@ -124,14 +124,18 @@ module Volt
     end
 
     # Assign multiple attributes as a hash, directly.
-    def assign_attributes(attrs, initial_setup = false)
-      @attributes = {}
+    def assign_attributes(attrs, initial_setup=false, skip_changes=false)
+      @attributes ||= {}
 
       attrs = wrap_values(attrs)
 
       if attrs
         # When doing a mass-assign, we don't validate or save until the end.
-        Model.no_change_tracking do
+        if initial_setup || skip_changes
+          Model.no_change_tracking do
+            assign_all_attributes(attrs, skip_changes)
+          end
+        else
           assign_all_attributes(attrs)
         end
       else
@@ -383,13 +387,14 @@ module Volt
     end
 
     # Used internally from other methods that assign all attributes
-    def assign_all_attributes(attrs)
+    def assign_all_attributes(attrs, track_changes=false)
       # Assign each attribute using setters
       attrs.each_pair do |key, value|
         key = key.to_sym
 
         # Track the change, since assign_all_attributes runs with no_change_tracking
-        attribute_will_change!(key, @attributes[key])
+        old_val = @attributes[key]
+        attribute_will_change!(key, old_val) if track_changes && old_val != value
 
         if self.respond_to?(:"#{key}=")
           # If a method without an underscore is defined, call that.