+++ spec/apps/kitchen_sink/app/main/config/routes.rb
@@ -1,16 +1,24 @@
 # See https://github.com/voltrb/volt#routes for more info on routes
 
-get '/bindings/{{ route_test }}', action: 'bindings'
-get '/bindings', action: 'bindings'
-get '/store', action: 'store'
-get '/cookie_test', action: 'cookie_test'
-get '/flash', action: 'flash'
-get '/yield', action: 'yield'
-get '/todos', controller: 'todos'
+client '/bindings/{{ route_test }}', action: 'bindings'
+client '/bindings', action: 'bindings'
+client '/store', action: 'store'
+client '/cookie_test', action: 'cookie_test'
+client '/flash', action: 'flash'
+client '/yield', action: 'yield'
+client '/todos', controller: 'todos'
 
 # Signup/login routes
-get '/signup', controller: 'user-templates', action: 'signup'
-get '/login', controller: 'user-templates', action: 'login'
+client '/signup', controller: 'user-templates', action: 'signup'
+client '/login', controller: 'user-templates', action: 'login'
+
+# HTTP endpoints
+get '/simple_http', controller: 'simple_http', action: 'index'
+get '/simple_http/store', controller: 'simple_http', action: 'show'
+post '/simple_http/upload', controller: 'simple_http', action: 'upload'
+
+# Route for file uploads
+client '/upload', controller: 'upload', action: 'index'
 
 # The main route, this should be last.  It will match any params not previously matched.
-get '/', {}
+client '/', {}